# docs/swagger.yaml
basePath: /
definitions:
  entity.AllSetting:
    properties:
      datepicker:
        description: Date picker format
        type: string
      expireDiff:
        description: Expiration warning threshold in days
        type: integer
      externalTrafficInformEnable:
        description: Enable external traffic reporting
        type: boolean
      externalTrafficInformURI:
        description: URI for external traffic reporting
        type: string
      ldapAutoCreate:
        type: boolean
      ldapAutoDelete:
        type: boolean
      ldapBaseDN:
        type: string
      ldapBindDN:
        type: string
      ldapDefaultExpiryDays:
        type: integer
      ldapDefaultLimitIP:
        type: integer
      ldapDefaultTotalGB:
        type: integer
      ldapEnable:
        description: LDAP settings
        type: boolean
      ldapFlagField:
        description: Generic flag configuration
        type: string
      ldapHost:
        type: string
      ldapInboundTags:
        type: string
      ldapInvertFlag:
        type: boolean
      ldapPassword:
        type: string
      ldapPort:
        type: integer
      ldapSyncCron:
        type: string
      ldapTruthyValues:
        type: string
      ldapUseTLS:
        type: boolean
      ldapUserAttr:
        description: e.g., mail or uid
        type: string
      ldapUserFilter:
        type: string
      ldapVlessField:
        type: string
      pageSize:
        description: UI settings
        type: integer
      remarkModel:
        description: Remark model pattern for inbounds
        type: string
      sessionMaxAge:
        description: Session maximum age in minutes
        type: integer
      subAnnounce:
        description: Subscription announce
        type: string
      subCertFile:
        description: SSL certificate file for subscription server
        type: string
      subDomain:
        description: Domain for subscription server validation
        type: string
      subEnable:
        description: Subscription server settings
        type: boolean
      subEnableRouting:
        description: Enable routing for subscription
        type: boolean
      subEncrypt:
        description: Encrypt subscription responses
        type: boolean
      subJsonEnable:
        description: Enable JSON subscription endpoint
        type: boolean
      subJsonFragment:
        description: JSON subscription fragment configuration
        type: string
      subJsonMux:
        description: JSON subscription mux configuration
        type: string
      subJsonNoises:
        description: JSON subscription noise configuration
        type: string
      subJsonPath:
        description: Path for JSON subscription endpoint
        type: string
      subJsonRules:
        type: string
      subJsonURI:
        description: JSON subscription server URI
        type: string
      subKeyFile:
        description: SSL private key file for subscription server
        type: string
      subListen:
        description: Subscription server listen IP
        type: string
      subPath:
        description: Base path for subscription URLs
        type: string
      subPort:
        description: Subscription server port
        type: integer
      subProfileUrl:
        description: Subscription profile URL
        type: string
      subRoutingRules:
        description: Subscription global routing rules (Only for Happ)
        type: string
      subShowInfo:
        description: Show client information in subscriptions
        type: boolean
      subSupportUrl:
        description: Subscription support URL
        type: string
      subTitle:
        description: Subscription title
        type: string
      subURI:
        description: Subscription server URI
        type: string
      subUpdates:
        description: Subscription update interval in minutes
        type: integer
      tgBotAPIServer:
        description: Custom API server for Telegram bot
        type: string
      tgBotBackup:
        description: Enable database backup via Telegram
        type: boolean
      tgBotChatId:
        description: Telegram chat ID for notifications
        type: string
      tgBotEnable:
        description: Telegram bot settings
        type: boolean
      tgBotLoginNotify:
        description: Send login notifications
        type: boolean
      tgBotProxy:
        description: Proxy URL for Telegram bot
        type: string
      tgBotToken:
        description: Telegram bot token
        type: string
      tgCpu:
        description: CPU usage threshold for alerts
        type: integer
      tgLang:
        description: Telegram bot language
        type: string
      tgRunTime:
        description: Cron schedule for Telegram notifications
        type: string
      timeLocation:
        description: Security settings
        type: string
      trafficDiff:
        description: Traffic warning threshold percentage
        type: integer
      twoFactorEnable:
        description: Enable two-factor authentication
        type: boolean
      twoFactorToken:
        description: Two-factor authentication token
        type: string
      webBasePath:
        description: Base path for web panel URLs
        type: string
      webCertFile:
        description: Path to SSL certificate file for web server
        type: string
      webDomain:
        description: Web server domain for domain validation
        type: string
      webKeyFile:
        description: Path to SSL private key file for web server
        type: string
      webListen:
        description: Web server settings
        type: string
      webPort:
        description: Web server port number
        type: integer
    type: object
  entity.Msg:
    properties:
      msg:
        description: Response message text
        type: string
      obj:
        description: Optional data object
      success:
        description: Indicates if the operation was successful
        type: boolean
    type: object
...
  model.Slave:
    properties:
      address:
        description: Slave IP or Domain
        type: string
      id:
        type: integer
      lastSeen:
        type: integer
      name:
        type: string
      port:
        description: Slave Port (optional if using reverse WS)
        type: integer
      secret:
        description: Auth Token for Slave
        type: string
      status:
        description: online, offline
        type: string
      systemStats:
        description: CPU/Mem stats (JSON)
        type: string
      version:
        description: Slave version
        type: string
    type: object
...
host: localhost:2053
info:
  contact: {}
  description: API documentation for 3X-UI Panel (Master-Slave Cluster)
  title: 3X-UI Cluster API
  version: "1.0"
paths:
...
  /panel/api/setting/update:
    post:
      consumes:
      - application/json
      description: Updates all panel settings
      parameters:
      - description: Settings data
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/entity.AllSetting'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Update settings
      tags:
      - Settings
...
  /panel/api/slave-certs/del/{id}:
    post:
      description: Deletes a TLS certificate by ID
      parameters:
      - description: Certificate ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Delete certificate
      tags:
      - SlaveCerts
...
  /panel/api/slave-certs/slave/{slaveId}:
    get:
      description: Returns TLS certificates for a specific slave node
      parameters:
      - description: Slave ID
        in: path
        name: slaveId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get certificates for slave
      tags:
      - SlaveCerts
  /panel/api/slave/add:
    post:
      consumes:
      - application/json
      description: Registers a new slave node
      parameters:
      - description: Slave data
        in: body
        name: slave
        required: true
        schema:
          $ref: '#/definitions/model.Slave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Add slave
      tags:
      - Slaves
  /panel/api/slave/connect:
    get:
      description: WebSocket endpoint for slave-to-master communication
      parameters:
      - description: Slave secret key
        in: query
        name: secret
        required: true
        type: string
      responses: {}
      summary: Connect slave (WebSocket)
      tags:
      - Slaves
  /panel/api/slave/del/{id}:
    post:
      description: Deletes a slave node with cascade deletion of all associated data
      parameters:
      - description: Slave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Delete slave
      tags:
      - Slaves
  /panel/api/slave/install/{id}:
    get:
      description: Generates a bash install command for setting up a slave node
      parameters:
      - description: Slave ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.Msg'
      summary: Get install command
      tags:
      - Slaves
...
securityDefinitions:
  SessionAuth:
    in: cookie
    name: 3x-ui
    type: apiKey
swagger: "2.0"